Remove util/heap.h dependency
