+++ src.mk
@@ -76,7 +76,7 @@ LIB_SOURCES =                                                   \
   db/memtable_list.cc                                           \
   db/merge_helper.cc                                            \
   db/merge_operator.cc                                          \
-  db/multi_cf_iterator.cc                                       \
+  db/multi_cf_iterator_impl.cc                                  \
   db/output_validator.cc                                        \
   db/periodic_task_scheduler.cc                                 \
   db/range_del_aggregator.cc                                    \
